refactor(pipeline): split effects.rs into focused sub-modules
effects.rs had grown past the 1500-line guideline (1756 lines) with the
crop/zoom work. Split it following the builder/ pattern — a pure move
with only import/visibility adjustments, no logic changes:

- effects/take.rs — the PGM take engine (trigger_transition)
- effects/mixer_ops.rs — vision mixer runtime ops (bus selection, DSK,
  FTB, overlay alpha, PiP config, input resolutions)
- effects/mixer_layout.rs — pad geometry/crop appliers shared by the
  take engine, mixer ops and the reactive geometry probe
- effects/misc.rs — block-generic effects (animate input, loudness
  reset, recorder split, thumbnail capture)
- effects/mod.rs — shared dist_canvas_size + crate-facing re-exports
