+++ backend/src/gst/pipeline/effects/misc.rs
@@ -1,0 +1,149 @@
+//! Block-generic live effects: input animation, loudness reset, recorder
+//! split, thumbnail capture.
+
+use super::super::{PipelineError, PipelineManager};
+use gstreamer::prelude::*;
+use tracing::info;
+
+impl PipelineManager {
+    /// Animate a single input's position/size on a compositor block.
+    #[allow(clippy::too_many_arguments)]
+    pub fn animate_input(
+        &self,
+        block_instance_id: &str,
+        input_index: usize,
+        target_xpos: Option<i32>,
+        target_ypos: Option<i32>,
+        target_width: Option<i32>,
+        target_height: Option<i32>,
+        duration_ms: u64,
+    ) -> Result<(), PipelineError> {
+        use crate::gst::transitions::TransitionController;
+
+        info!(
+            "Animating input {} on {} to ({:?}, {:?}, {:?}, {:?}) over {}ms",
+            input_index,
+            block_instance_id,
+            target_xpos,
+            target_ypos,
+            target_width,
+            target_height,
+            duration_ms
+        );
+
+        // Find the mixer element for this block
+        let mixer_id = format!("{}:mixer", block_instance_id);
+        let mixer = self
+            .elements
+            .get(&mixer_id)
+            .ok_or_else(|| PipelineError::ElementNotFound(mixer_id.clone()))?;
+
+        let (canvas_width, canvas_height) = self.dist_canvas_size(block_instance_id);
+
+        // Create transition controller and animate
+        let controller = TransitionController::new(mixer.clone(), canvas_width, canvas_height);
+        controller
+            .animate_input(
+                input_index,
+                target_xpos,
+                target_ypos,
+                target_width,
+                target_height,
+                duration_ms,
+                &self.pipeline,
+            )
+            .map_err(|e| PipelineError::TransitionError(e.to_string()))?;
+
+        Ok(())
+    }
+
+    /// Reset accumulated loudness measurements on an EBU R128 meter block.
+    pub fn reset_loudness(&self, block_instance_id: &str) -> Result<(), PipelineError> {
+        let element_id = format!("{}:ebur128level", block_instance_id);
+        let element = self
+            .elements
+            .get(&element_id)
+            .ok_or_else(|| PipelineError::ElementNotFound(element_id.clone()))?;
+        element.emit_by_name::<()>("reset", &[]);
+        info!("Reset loudness measurements on {}", block_instance_id);
+        Ok(())
+    }
+
+    /// Force an immediate file split on a recorder block.
+    ///
+    /// Emits the `split-now` signal on the splitmuxsink element, which triggers
+    /// a file split at the next keyframe boundary.
+    pub fn recorder_split_now(&self, block_instance_id: &str) -> Result<(), PipelineError> {
+        use crate::blocks::builtin::recorder::SPLITMUXSINK_SUFFIX;
+        let element_id = format!("{}:{}", block_instance_id, SPLITMUXSINK_SUFFIX);
+        let element = self.elements.get(&element_id).ok_or_else(|| {
+            PipelineError::ElementNotFound(format!(
+                "{} (is this a recorder block in ts_passthrough mode?)",
+                element_id
+            ))
+        })?;
+        element.emit_by_name::<()>("split-now", &[]);
+        info!(
+            "Triggered split-now on recorder block {}",
+            block_instance_id
+        );
+        Ok(())
+    }
+
+    /// Capture a thumbnail from a block's tee element at the given index.
+    ///
+    /// Lazily attaches a GStreamer-native processing branch to the block's tee
+    /// element. The branch does format conversion and scaling using GStreamer
+    /// elements, with lightweight JPEG encoding in the appsink callback.
+    ///
+    /// The meaning of `index` depends on the block type:
+    /// - **Compositor**: input index (each input has its own tee named `{block_id}:thumb_tee_{index}`)
+    /// - **Thumbnail block**: always 0 (single tee named `{block_id}:tee`)
+    pub fn capture_block_thumbnail(
+        &self,
+        block_id: &str,
+        index: usize,
+    ) -> Result<Vec<u8>, PipelineError> {
+        use crate::gst::thumbnail_tap::{ThumbnailTap, ThumbnailTapConfig};
+
+        let mut taps = self.thumbnail_taps.lock().unwrap();
+        let block_taps = taps.entry(block_id.to_string()).or_default();
+
+        // Ensure we have a tap for this index (lazy creation)
+        while block_taps.len() <= index {
+            let idx = block_taps.len();
+            // Try compositor naming first ({block_id}:thumb_tee_{idx}),
+            // fall back to simple naming ({block_id}:tee) for index 0.
+            let tee_name = format!("{}:thumb_tee_{}", block_id, idx);
+            let tee = self
+                .pipeline
+                .by_name(&tee_name)
+                .or_else(|| {
+                    if idx == 0 {
+                        self.pipeline.by_name(&format!("{}:tee", block_id))
+                    } else {
+                        None
+                    }
+                })
+                .ok_or_else(|| {
+                    PipelineError::ElementNotFound(format!(
+                        "Thumbnail tee not found: {} (block {})",
+                        tee_name, block_id
+                    ))
+                })?;
+
+            let name_prefix = format!("{}:thumb_{}", block_id, idx);
+            let tap = ThumbnailTap::new_with_tee(
+                &self.pipeline,
+                &name_prefix,
+                tee,
+                ThumbnailTapConfig::default(),
+            );
+            block_taps.push(tap);
+        }
+
+        block_taps[index]
+            .get_thumbnail()
+            .map_err(|e| PipelineError::ThumbnailCapture(e.to_string()))
+    }
+}
