+++ backend/src/gst/pipeline/effects/mixer_ops.rs
@@ -1,576 +1,19 @@
-use super::{PipelineError, PipelineManager};
+//! Vision mixer runtime operations: PVW/PGM bus selection, DSK toggles,
+//! fade-to-black, multiview overlay alpha, live PiP (re)configuration and
+//! negotiated input resolutions.
+
+use super::super::{PipelineError, PipelineManager};
 use gstreamer as gst;
 use gstreamer::prelude::*;
-use tracing::{debug, info, warn};
+use tracing::{info, warn};
+
+use crate::gst::crop::set_pad_crop;
+
+use super::mixer_layout::{
+    apply_pip_crop_after_morph, apply_pip_layout_to_region, find_pad, pads_for_source,
+};
 
 impl PipelineManager {
-    /// Get the distribution compositor canvas size from its capsfilter.
-    fn dist_canvas_size(&self, block_instance_id: &str) -> (i32, i32) {
-        let default =
-            strom_types::parse_resolution_string(strom_types::vision_mixer::DEFAULT_PGM_RESOLUTION)
-                .map(|(w, h)| (w as i32, h as i32))
-                .expect("DEFAULT_PGM_RESOLUTION must be valid");
-
-        let capsfilter_id = format!("{}:capsfilter_dist", block_instance_id);
-        self.elements
-            .get(&capsfilter_id)
-            .and_then(|cf| cf.property::<Option<gst::Caps>>("caps"))
-            .and_then(|caps| {
-                let s = caps.structure(0)?;
-                Some((
-                    s.get::<i32>("width").unwrap_or(default.0),
-                    s.get::<i32>("height").unwrap_or(default.1),
-                ))
-            })
-            .unwrap_or(default)
-    }
-
-    /// Trigger a transition on a compositor/mixer block.
-    ///
-    /// Reads the authoritative PGM/PVW source from overlay state. When either
-    /// bus is a PiP, takes the source-aware path that animates pads across
-    /// dist + multiview compositors; otherwise runs a standard single-pad
-    /// transition on the mv mixer.
-    ///
-    /// Returns `(was_ftb_cancelled, old_pgm, new_pgm, actual_kind)`. The two
-    /// middle elements are `None` when the corresponding bus is a PiP source.
-    /// `actual_kind` is the transition that actually ran — differs from
-    /// `transition_type` when the engine downgraded the request (e.g. Slide
-    /// across heterogeneous PiP/input sources downgrades to "fade").
-    pub fn trigger_transition(
-        &self,
-        block_instance_id: &str,
-        from_input: usize,
-        to_input: usize,
-        transition_type: &str,
-        duration_ms: u64,
-    ) -> Result<(bool, Option<usize>, Option<usize>, String), PipelineError> {
-        use crate::gst::transitions::{TransitionController, TransitionType};
-
-        debug!(
-            "Triggering {} transition on {} from input {} to {} ({}ms)",
-            transition_type, block_instance_id, from_input, to_input, duration_ms
-        );
-
-        // Find the mixer element for this block
-        let mixer_id = format!("{}:mixer", block_instance_id);
-        let mixer = self
-            .elements
-            .get(&mixer_id)
-            .ok_or_else(|| PipelineError::ElementNotFound(mixer_id.clone()))?;
-
-        // Read authoritative PGM/PVW source from overlay state. `None` means
-        // the bus is a PiP — handled by the PiP-aware branch below.
-        let overlay_state =
-            crate::blocks::builtin::vision_mixer::overlay::get_overlay_state(block_instance_id);
-        let num_video_inputs = overlay_state
-            .as_ref()
-            .map(|s| s.num_inputs)
-            .unwrap_or(usize::MAX);
-        // When overlay state is missing entirely, fall back to the request-
-        // provided indices. When state exists but reports None (PiP on bus),
-        // pass None through so the PiP-aware branch below handles it.
-        let old_pgm: Option<usize> = match overlay_state.as_ref() {
-            Some(s) => s.pgm_input(),
-            None => Some(from_input),
-        };
-        let new_pgm: Option<usize> = match overlay_state.as_ref() {
-            Some(s) => s.pvw_input(),
-            None => Some(to_input),
-        };
-
-        // Auto-cancel FTB if active
-        let was_ftb = overlay_state
-            .as_ref()
-            .map(|s| {
-                s.ftb_active
-                    .swap(false, std::sync::atomic::Ordering::Relaxed)
-            })
-            .unwrap_or(false);
-        if was_ftb {
-            info!(
-                "Auto-cancelling FTB before transition on {}",
-                block_instance_id
-            );
-        }
-
-        let (canvas_width, canvas_height) = self.dist_canvas_size(block_instance_id);
-
-        // --- PiP-aware Take ---
-        // If either bus is currently a PiP (or will become one via swap), use a
-        // Source-aware path. Cut snaps geometry+alpha; Fade cross-fades alphas
-        // across both compositor regions (dist for PGM, mv PVW for PVW).
-        // Slide/Push/Dip-to-Black with PiP downgrade to Fade — explicit position
-        // animation across heterogeneous Source kinds isn't supported yet.
-        if let Some(state) = overlay_state.as_ref() {
-            let old_pgm_pip = state.pgm_pip();
-            let old_pvw_pip = state.pvw_pip();
-            if old_pgm_pip.is_some() || old_pvw_pip.is_some() {
-                let mv_comp_id = format!("{}:mv_comp", block_instance_id);
-                let mv_comp = self
-                    .elements
-                    .get(&mv_comp_id)
-                    .ok_or_else(|| PipelineError::ElementNotFound(mv_comp_id.clone()))?;
-
-                let parsed = transition_type.parse::<TransitionType>().ok();
-                let is_cut = duration_ms == 0 || matches!(parsed, Some(TransitionType::Cut));
-                // Explicit position animation across heterogeneous Source kinds
-                // (input ↔ PiP) isn't supported yet — Slide/Push/Dip-to-Black
-                // silently degrade to Fade in this branch. Surface that in the
-                // log so operators don't think the requested transition ran.
-                if !is_cut
-                    && !matches!(
-                        parsed,
-                        Some(TransitionType::Fade) | Some(TransitionType::Cut)
-                    )
-                {
-                    info!(
-                        "PiP-aware Take on {}: transition '{}' downgraded to Fade ({}ms) — non-Fade transitions across PiPs not supported yet",
-                        block_instance_id, transition_type, duration_ms
-                    );
-                }
-
-                // Swap Source state PVW ↔ PGM.
-                let new_pgm_pip = old_pvw_pip;
-                let new_pvw_pip = old_pgm_pip;
-                let new_pgm_swap = state.pvw_input();
-                let new_pvw_swap = state.pgm_input();
-
-                let dist_region = (0, 0, canvas_width, canvas_height);
-                let r = &state.layout.pvw_rect;
-                let pvw_region = (r.x as i32, r.y as i32, r.w as i32, r.h as i32);
-                // Fallback aspect = PGM canvas aspect (typically 16:9), used
-                // for tile-grid cells and for inputs whose caps are unknown.
-                let src_aspect = if canvas_height > 0 {
-                    canvas_width as f64 / canvas_height as f64
-                } else {
-                    16.0 / 9.0
-                };
-                let src_aspects =
-                    self.vision_mixer_source_aspects(block_instance_id, state.num_inputs);
-
-                let old_dist_targets = pads_for_source(
-                    state,
-                    old_pgm_pip,
-                    old_pgm,
-                    dist_region,
-                    0,
-                    strom_types::vision_mixer::DIST_PGM_ZORDER,
-                    strom_types::vision_mixer::DIST_PIP_OVERLAY_ZORDER,
-                    src_aspect,
-                    &src_aspects,
-                );
-                let new_dist_targets = pads_for_source(
-                    state,
-                    new_pgm_pip,
-                    new_pgm_swap,
-                    dist_region,
-                    0,
-                    strom_types::vision_mixer::DIST_PGM_ZORDER,
-                    strom_types::vision_mixer::DIST_PIP_OVERLAY_ZORDER,
-                    src_aspect,
-                    &src_aspects,
-                );
-                let old_pvw_targets = pads_for_source(
-                    state,
-                    old_pvw_pip,
-                    state.pvw_input(),
-                    pvw_region,
-                    state.num_inputs + 1,
-                    strom_types::vision_mixer::MV_BIG_DISPLAY_ZORDER,
-                    strom_types::vision_mixer::MV_PVW_PIP_OVERLAY_ZORDER,
-                    src_aspect,
-                    &src_aspects,
-                );
-                let new_pvw_targets = pads_for_source(
-                    state,
-                    new_pvw_pip,
-                    new_pvw_swap,
-                    pvw_region,
-                    state.num_inputs + 1,
-                    strom_types::vision_mixer::MV_BIG_DISPLAY_ZORDER,
-                    strom_types::vision_mixer::MV_PVW_PIP_OVERLAY_ZORDER,
-                    src_aspect,
-                    &src_aspects,
-                );
-
-                if is_cut {
-                    // Snap apply: hide everything in the region, then set new active pads.
-                    if let Some(p) = new_pgm_pip {
-                        apply_pip_layout_to_region(
-                            mixer,
-                            0,
-                            state.num_inputs,
-                            dist_region,
-                            state.pip_bg_input(p),
-                            &state.pip_zones(p),
-                            &state.pip_transforms(p),
-                            strom_types::vision_mixer::DIST_PGM_ZORDER,
-                            strom_types::vision_mixer::DIST_PIP_OVERLAY_ZORDER,
-                            src_aspect,
-                            &src_aspects,
-                        );
-                    } else {
-                        apply_input_group_to_region(
-                            mixer,
-                            0,
-                            state.num_inputs,
-                            dist_region,
-                            new_pgm_swap,
-                            strom_types::vision_mixer::DIST_PGM_ZORDER,
-                            src_aspect,
-                            &src_aspects,
-                        );
-                    }
-                    if let Some(p) = new_pvw_pip {
-                        apply_pip_layout_to_region(
-                            mv_comp,
-                            state.num_inputs + 1,
-                            state.num_inputs,
-                            pvw_region,
-                            state.pip_bg_input(p),
-                            &state.pip_zones(p),
-                            &state.pip_transforms(p),
-                            strom_types::vision_mixer::MV_BIG_DISPLAY_ZORDER,
-                            strom_types::vision_mixer::MV_PVW_PIP_OVERLAY_ZORDER,
-                            src_aspect,
-                            &src_aspects,
-                        );
-                    } else {
-                        apply_input_group_to_region(
-                            mv_comp,
-                            state.num_inputs + 1,
-                            state.num_inputs,
-                            pvw_region,
-                            new_pvw_swap,
-                            strom_types::vision_mixer::MV_BIG_DISPLAY_ZORDER,
-                            src_aspect,
-                            &src_aspects,
-                        );
-                    }
-                } else {
-                    // Morph transition: sources shared between old and new
-                    // animate position+size; fresh sources alpha-fade in; departing
-                    // sources alpha-fade out. Runs in parallel on both dist (PGM)
-                    // and mv PVW big regions.
-                    let dist_controller =
-                        TransitionController::new(mixer.clone(), canvas_width, canvas_height);
-                    dist_controller
-                        .animate_pad_transition(
-                            &old_dist_targets,
-                            &new_dist_targets,
-                            duration_ms,
-                            &self.pipeline,
-                        )
-                        .map_err(|e| PipelineError::TransitionError(e.to_string()))?;
-
-                    let mv_controller = TransitionController::new(
-                        mv_comp.clone(),
-                        state.layout.canvas_width as i32,
-                        state.layout.canvas_height as i32,
-                    );
-                    mv_controller
-                        .animate_pad_transition(
-                            &old_pvw_targets,
-                            &new_pvw_targets,
-                            duration_ms,
-                            &self.pipeline,
-                        )
-                        .map_err(|e| PipelineError::TransitionError(e.to_string()))?;
-
-                    // Apply per-source crop for the incoming compositions:
-                    // pads shared between old and new (e.g. a source that is
-                    // fullscreen on PGM and cropped in the incoming PiP)
-                    // animate their crop in sync with the geometry morph;
-                    // fresh pads snap; fading-out pads keep their crop until
-                    // hidden. An input-mode side uses an empty transform map,
-                    // so shared pads animate their punch-out back to no crop.
-                    let new_pgm_transforms = new_pgm_pip
-                        .map(|p| state.pip_transforms(p))
-                        .unwrap_or_default();
-                    apply_pip_crop_after_morph(
-                        mixer,
-                        &dist_controller,
-                        &self.pipeline,
-                        &new_pgm_transforms,
-                        state.num_inputs,
-                        0,
-                        &old_dist_targets,
-                        &new_dist_targets,
-                        duration_ms,
-                    );
-                    let new_pvw_transforms = new_pvw_pip
-                        .map(|p| state.pip_transforms(p))
-                        .unwrap_or_default();
-                    apply_pip_crop_after_morph(
-                        mv_comp,
-                        &mv_controller,
-                        &self.pipeline,
-                        &new_pvw_transforms,
-                        state.num_inputs,
-                        state.num_inputs + 1,
-                        &old_pvw_targets,
-                        &new_pvw_targets,
-                        duration_ms,
-                    );
-                }
-
-                // Persist new state.
-                state.set_pgm_pip(new_pgm_pip);
-                state.set_pvw_pip(new_pvw_pip);
-                state.set_pgm_input(new_pgm_swap);
-                state.set_pvw_input(new_pvw_swap);
-                crate::blocks::builtin::vision_mixer::overlay::trigger_overlay_update(
-                    block_instance_id,
-                );
-
-                info!(
-                    "PiP-aware Take on {} ({}{}ms): PGM pip={:?} input={:?}; PVW pip={:?} input={:?}",
-                    block_instance_id,
-                    if is_cut { "Cut, " } else { "Fade, " },
-                    duration_ms,
-                    new_pgm_pip,
-                    new_pgm_swap,
-                    new_pvw_pip,
-                    new_pvw_swap,
-                );
-                // After the downgrade guard above, the PiP-aware branch only
-                // ever runs a Cut or a Fade.
-                let actual_kind = if is_cut { "cut" } else { "fade" }.to_string();
-                return Ok((was_ftb, old_pgm, new_pgm_swap, actual_kind));
-            }
-        }
-
-        // Reset all video pads to a clean state before the transition:
-        // clear control bindings, restore alpha/position/size for the current
-        // PGM input. Single-input PGM only (multi-source compositions are PiPs).
-        let active_pgm_input = old_pgm;
-        let classic_aspects = self.vision_mixer_source_aspects(
-            block_instance_id,
-            if num_video_inputs == usize::MAX {
-                0
-            } else {
-                num_video_inputs
-            },
-        );
-        let canvas_aspect = if canvas_height > 0 {
-            canvas_width as f64 / canvas_height as f64
-        } else {
-            16.0 / 9.0
-        };
-
-        for pad in mixer.sink_pads() {
-            let name = pad.name();
-            if name.starts_with("sink_") {
-                if let Ok(idx) = name.trim_start_matches("sink_").parse::<usize>() {
-                    for prop in ["alpha", "xpos", "ypos", "width", "height"] {
-                        if let Some(binding) = pad.control_binding(prop) {
-                            pad.remove_control_binding(&binding);
-                        }
-                    }
-                    if idx < num_video_inputs {
-                        // Classic takes are input↔input — wipe any crop left
-                        // behind by an earlier PiP render on these pads.
-                        set_pad_crop(&pad, &Default::default());
-                        // Explicit geometry: aspect-fit the source into the
-                        // canvas (pads run sizing-policy=none).
-                        let aspect = classic_aspects.get(&idx).copied().unwrap_or(canvas_aspect);
-                        let (x, y, w, h) = strom_types::vision_mixer::aspect_fit_rect(
-                            0,
-                            0,
-                            canvas_width,
-                            canvas_height,
-                            aspect,
-                        );
-                        pad.set_property("xpos", x);
-                        pad.set_property("ypos", y);
-                        pad.set_property("width", w);
-                        pad.set_property("height", h);
-                        if Some(idx) == active_pgm_input {
-                            pad.set_property("alpha", 1.0f64);
-                            pad.set_property("zorder", strom_types::vision_mixer::DIST_PGM_ZORDER);
-                        } else {
-                            pad.set_property("alpha", 0.0f64);
-                        }
-                    } else if let Some(state) = overlay_state.as_ref() {
-                        let dsk_idx = idx - num_video_inputs;
-                        let enabled = dsk_idx < state.dsk_enabled.len()
-                            && state.dsk_enabled[dsk_idx]
-                                .load(std::sync::atomic::Ordering::Relaxed);
-                        let alpha = if enabled { 1.0f64 } else { 0.0f64 };
-                        pad.set_property("alpha", alpha);
-                    }
-                }
-            }
-        }
-
-        // Parse transition type
-        let trans_type = transition_type.parse::<TransitionType>().map_err(|_| {
-            PipelineError::InvalidProperty {
-                element: block_instance_id.to_string(),
-                property: "transition_type".to_string(),
-                reason: format!("Unknown transition type: {}", transition_type),
-            }
-        })?;
-
-        // Single-input transition only. Multi-source compositions are now
-        // expressed as PiPs which are handled by the PiP-aware branch above.
-        let from = old_pgm.unwrap_or(from_input);
-        let to = new_pgm.unwrap_or(to_input);
-        let controller = TransitionController::new(mixer.clone(), canvas_width, canvas_height);
-        controller
-            .transition(from, to, trans_type, duration_ms, &self.pipeline)
-            .map_err(|e| PipelineError::TransitionError(e.to_string()))?;
-
-        let actual_kind = if duration_ms == 0 {
-            TransitionType::Cut.to_string()
-        } else {
-            trans_type.to_string()
-        };
-        Ok((was_ftb, old_pgm, new_pgm, actual_kind))
-    }
-
-    /// Animate a single input's position/size on a compositor block.
-    #[allow(clippy::too_many_arguments)]
-    pub fn animate_input(
-        &self,
-        block_instance_id: &str,
-        input_index: usize,
-        target_xpos: Option<i32>,
-        target_ypos: Option<i32>,
-        target_width: Option<i32>,
-        target_height: Option<i32>,
-        duration_ms: u64,
-    ) -> Result<(), PipelineError> {
-        use crate::gst::transitions::TransitionController;
-
-        info!(
-            "Animating input {} on {} to ({:?}, {:?}, {:?}, {:?}) over {}ms",
-            input_index,
-            block_instance_id,
-            target_xpos,
-            target_ypos,
-            target_width,
-            target_height,
-            duration_ms
-        );
-
-        // Find the mixer element for this block
-        let mixer_id = format!("{}:mixer", block_instance_id);
-        let mixer = self
-            .elements
-            .get(&mixer_id)
-            .ok_or_else(|| PipelineError::ElementNotFound(mixer_id.clone()))?;
-
-        let (canvas_width, canvas_height) = self.dist_canvas_size(block_instance_id);
-
-        // Create transition controller and animate
-        let controller = TransitionController::new(mixer.clone(), canvas_width, canvas_height);
-        controller
-            .animate_input(
-                input_index,
-                target_xpos,
-                target_ypos,
-                target_width,
-                target_height,
-                duration_ms,
-                &self.pipeline,
-            )
-            .map_err(|e| PipelineError::TransitionError(e.to_string()))?;
-
-        Ok(())
-    }
-
-    /// Reset accumulated loudness measurements on an EBU R128 meter block.
-    pub fn reset_loudness(&self, block_instance_id: &str) -> Result<(), PipelineError> {
-        let element_id = format!("{}:ebur128level", block_instance_id);
-        let element = self
-            .elements
-            .get(&element_id)
-            .ok_or_else(|| PipelineError::ElementNotFound(element_id.clone()))?;
-        element.emit_by_name::<()>("reset", &[]);
-        info!("Reset loudness measurements on {}", block_instance_id);
-        Ok(())
-    }
-
-    /// Force an immediate file split on a recorder block.
-    ///
-    /// Emits the `split-now` signal on the splitmuxsink element, which triggers
-    /// a file split at the next keyframe boundary.
-    pub fn recorder_split_now(&self, block_instance_id: &str) -> Result<(), PipelineError> {
-        use crate::blocks::builtin::recorder::SPLITMUXSINK_SUFFIX;
-        let element_id = format!("{}:{}", block_instance_id, SPLITMUXSINK_SUFFIX);
-        let element = self.elements.get(&element_id).ok_or_else(|| {
-            PipelineError::ElementNotFound(format!(
-                "{} (is this a recorder block in ts_passthrough mode?)",
-                element_id
-            ))
-        })?;
-        element.emit_by_name::<()>("split-now", &[]);
-        info!(
-            "Triggered split-now on recorder block {}",
-            block_instance_id
-        );
-        Ok(())
-    }
-
-    /// Capture a thumbnail from a block's tee element at the given index.
-    ///
-    /// Lazily attaches a GStreamer-native processing branch to the block's tee
-    /// element. The branch does format conversion and scaling using GStreamer
-    /// elements, with lightweight JPEG encoding in the appsink callback.
-    ///
-    /// The meaning of `index` depends on the block type:
-    /// - **Compositor**: input index (each input has its own tee named `{block_id}:thumb_tee_{index}`)
-    /// - **Thumbnail block**: always 0 (single tee named `{block_id}:tee`)
-    pub fn capture_block_thumbnail(
-        &self,
-        block_id: &str,
-        index: usize,
-    ) -> Result<Vec<u8>, PipelineError> {
-        use crate::gst::thumbnail_tap::{ThumbnailTap, ThumbnailTapConfig};
-
-        let mut taps = self.thumbnail_taps.lock().unwrap();
-        let block_taps = taps.entry(block_id.to_string()).or_default();
-
-        // Ensure we have a tap for this index (lazy creation)
-        while block_taps.len() <= index {
-            let idx = block_taps.len();
-            // Try compositor naming first ({block_id}:thumb_tee_{idx}),
-            // fall back to simple naming ({block_id}:tee) for index 0.
-            let tee_name = format!("{}:thumb_tee_{}", block_id, idx);
-            let tee = self
-                .pipeline
-                .by_name(&tee_name)
-                .or_else(|| {
-                    if idx == 0 {
-                        self.pipeline.by_name(&format!("{}:tee", block_id))
-                    } else {
-                        None
-                    }
-                })
-                .ok_or_else(|| {
-                    PipelineError::ElementNotFound(format!(
-                        "Thumbnail tee not found: {} (block {})",
-                        tee_name, block_id
-                    ))
-                })?;
-
-            let name_prefix = format!("{}:thumb_{}", block_id, idx);
-            let tap = ThumbnailTap::new_with_tee(
-                &self.pipeline,
-                &name_prefix,
-                tee,
-                ThumbnailTapConfig::default(),
-            );
-            block_taps.push(tap);
-        }
-
-        block_taps[index]
-            .get_thumbnail()
-            .map_err(|e| PipelineError::ThumbnailCapture(e.to_string()))
-    }
-
     /// Select a preview input on a vision mixer block.
     ///
     /// Replaces the PVW source with `input` (clearing any PiP-on-PVW mode).
@@ -1483,274 +926,5 @@ impl PipelineManager {
             block_instance_id, pip_idx, bg, zones
         );
         Ok(())
-    }
-}
-
-/// Find a pad by name on an element, checking both static and request pads.
-/// `static_pad()` doesn't find request pads on aggregator elements like glvideomixer.
-pub(crate) fn find_pad(element: &gst::Element, pad_name: &str) -> Option<gst::Pad> {
-    element.static_pad(pad_name).or_else(|| {
-        element
-            .pads()
-            .into_iter()
-            .find(|p| p.name().as_str() == pad_name)
-    })
-}
-
-use crate::gst::crop::set_pad_crop;
-
-/// Apply per-source crop to a region after a zone morph: pads present in both
-/// old and new layouts animate crop alongside the geometry morph; pads about
-/// to be revealed snap to their crop; fading-out pads keep theirs until hidden.
-#[allow(clippy::too_many_arguments)]
-fn apply_pip_crop_after_morph(
-    compositor: &gst::Element,
-    ctrl: &crate::gst::transitions::TransitionController,
-    pipeline: &gst::Pipeline,
-    transforms: &strom_types::vision_mixer::PipTransforms,
-    num_inputs: usize,
-    pad_base: usize,
-    old: &[crate::gst::transitions::PadTarget],
-    new: &[crate::gst::transitions::PadTarget],
-    duration_ms: u64,
-) {
-    let old_set: std::collections::HashSet<usize> = old.iter().map(|t| t.pad_idx).collect();
-    let new_set: std::collections::HashSet<usize> = new.iter().map(|t| t.pad_idx).collect();
-    for i in 0..num_inputs {
-        let pad_idx = pad_base + i;
-        let crop = transforms.get(&i).copied().unwrap_or_default();
-        if old_set.contains(&pad_idx) && new_set.contains(&pad_idx) {
-            // Staying pad — ease the crop in sync with the geometry morph
-            // ("punch-in" zoom). No-op on the CPU backend.
-            if let Err(e) = ctrl.animate_pad_crop(pad_idx, &crop, duration_ms, pipeline) {
-                warn!("Crop morph failed on pad {}: {}", pad_idx, e);
-            }
-        } else if old_set.contains(&pad_idx) {
-            // Departing pad — keeps its crop while fading out; whichever path
-            // reveals it next re-applies the correct crop.
-        } else if let Some(pad) = find_pad(compositor, &format!("sink_{}", pad_idx)) {
-            set_pad_crop(&pad, &crop);
-        }
-    }
-}
-
-/// Compute per-pad targets `(pad_idx, x, y, w, h, zorder)` for a Source rendered
-/// into a compositor region. `pad_base` is the index offset of the first input
-/// pad in the region (e.g. 0 for dist_comp, num_inputs+1 for mv_comp PVW big).
-///
-/// Geometry is explicit (pads run `sizing-policy=none`): every rect is
-/// aspect-fitted with the source's effective aspect (crop-adjusted), so an
-/// odd-aspect source letterboxes correctly and a locked crop fills its box.
-#[allow(clippy::too_many_arguments)]
-fn pads_for_source(
-    state: &crate::blocks::builtin::vision_mixer::overlay::VisionMixerOverlayState,
-    pip: Option<usize>,
-    input: Option<usize>,
-    region: (i32, i32, i32, i32),
-    pad_base: usize,
-    bg_zorder: u32,
-    overlay_zorder: u32,
-    fallback_aspect: f64,
-    src_aspects: &strom_types::vision_mixer::SourceAspects,
-) -> Vec<crate::gst::transitions::PadTarget> {
-    use crate::gst::transitions::PadTarget;
-    use strom_types::vision_mixer::{aspect_fit_rect, effective_source_aspect};
-    let (rx, ry, rw, rh) = region;
-    if let Some(p) = pip {
-        let bg = state.pip_bg_input(p);
-        let zones = state.pip_zones(p);
-        let transforms = state.pip_transforms(p);
-        let layouts = strom_types::vision_mixer::resolve_zone_pads(
-            rx,
-            ry,
-            rw,
-            rh,
-            &zones,
-            fallback_aspect,
-            &transforms,
-            src_aspects,
-        );
-        // Defensive dedupe: `apply_vision_mixer_pip_config` already filters bg
-        // out of zone sources, but the on-disk state could become inconsistent
-        // (e.g. legacy config or future code that bypasses the API). Two
-        // PadTargets with the same `pad_idx` would confuse the transition
-        // planner — keep the bg entry, drop any overlapping zone source.
-        let mut out = Vec::new();
-        let mut seen_pad_idxs = std::collections::HashSet::new();
-        if let Some(b) = bg {
-            let aspect = effective_source_aspect(
-                src_aspects.get(&b).copied().unwrap_or(fallback_aspect),
-                transforms.get(&b),
-            );
-            let (x, y, w, h) = aspect_fit_rect(rx, ry, rw, rh, aspect);
-            out.push(PadTarget {
-                pad_idx: pad_base + b,
-                x,
-                y,
-                w,
-                h,
-                zorder: bg_zorder,
-            });
-            seen_pad_idxs.insert(pad_base + b);
-        }
-        for l in &layouts {
-            let pad_idx = pad_base + l.input;
-            if !seen_pad_idxs.insert(pad_idx) {
-                continue;
-            }
-            out.push(PadTarget {
-                pad_idx,
-                x: l.x,
-                y: l.y,
-                w: l.w,
-                h: l.h,
-                zorder: overlay_zorder + l.zorder_offset,
-            });
-        }
-        out
-    } else {
-        // Single-input source (multi-source compositions are PiPs now).
-        // Plain inputs never crop — fit with the raw source aspect.
-        input
-            .map(|idx| {
-                let aspect = src_aspects.get(&idx).copied().unwrap_or(fallback_aspect);
-                let (x, y, w, h) = aspect_fit_rect(rx, ry, rw, rh, aspect);
-                PadTarget {
-                    pad_idx: pad_base + idx,
-                    x,
-                    y,
-                    w,
-                    h,
-                    zorder: bg_zorder,
-                }
-            })
-            .into_iter()
-            .collect()
-    }
-}
-
-/// Apply a single-input source layout (aspect-fitted into the region) to a
-/// contiguous range of compositor sink pads. Pads not equal to the active
-/// input are hidden.
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn apply_input_group_to_region(
-    compositor: &gst::Element,
-    pad_base: usize,
-    num_inputs: usize,
-    region: (i32, i32, i32, i32),
-    active: Option<usize>,
-    fg_zorder: u32,
-    fallback_aspect: f64,
-    src_aspects: &strom_types::vision_mixer::SourceAspects,
-) {
-    let (rx, ry, rw, rh) = region;
-    for i in 0..num_inputs {
-        let pad_name = format!("sink_{}", pad_base + i);
-        let Some(pad) = find_pad(compositor, &pad_name) else {
-            continue;
-        };
-        // Plain-input mode never crops — wipe any crop left by a PiP render.
-        set_pad_crop(&pad, &Default::default());
-        // Clear any lingering control bindings from a previous morph/fade so
-        // our set_property writes aren't silently overridden by stale values.
-        for prop in ["alpha", "xpos", "ypos", "width", "height", "zorder"] {
-            if let Some(binding) = pad.control_binding(prop) {
-                pad.remove_control_binding(&binding);
-            }
-        }
-        if Some(i) == active {
-            let aspect = src_aspects.get(&i).copied().unwrap_or(fallback_aspect);
-            let (x, y, w, h) = strom_types::vision_mixer::aspect_fit_rect(rx, ry, rw, rh, aspect);
-            pad.set_property("xpos", x);
-            pad.set_property("ypos", y);
-            pad.set_property("width", w);
-            pad.set_property("height", h);
-            pad.set_property("alpha", 1.0f64);
-            pad.set_property("zorder", fg_zorder);
-        } else {
-            pad.set_property("alpha", 0.0f64);
-        }
-    }
-}
-
-/// Apply a PiP composition (bg fills the region, overlays auto-tile on top) to a
-/// contiguous range of compositor sink pads named `sink_{pad_base..pad_base+num_inputs}`.
-///
-/// Each pad whose input index matches `bg` becomes the background; each pad whose
-/// input is found in `overlays` is positioned at its corresponding tile; all other
-/// pads in the range are hidden (alpha=0).
-#[allow(clippy::too_many_arguments)]
-pub(crate) fn apply_pip_layout_to_region(
-    compositor: &gst::Element,
-    pad_base: usize,
-    num_inputs: usize,
-    region: (i32, i32, i32, i32),
-    bg: Option<usize>,
-    zones: &[strom_types::vision_mixer::Zone],
-    transforms: &strom_types::vision_mixer::PipTransforms,
-    bg_zorder: u32,
-    overlay_zorder: u32,
-    fallback_aspect: f64,
-    src_aspects: &strom_types::vision_mixer::SourceAspects,
-) {
-    use strom_types::vision_mixer::{aspect_fit_rect, effective_source_aspect};
-    let (rx, ry, rw, rh) = region;
-    let layouts = strom_types::vision_mixer::resolve_zone_pads(
-        rx,
-        ry,
-        rw,
-        rh,
-        zones,
-        fallback_aspect,
-        transforms,
-        src_aspects,
-    );
-    // Fast lookup: input → (x, y, w, h, zorder_offset).
-    let layout_map: std::collections::HashMap<usize, (i32, i32, i32, i32, u32)> = layouts
-        .iter()
-        .map(|l| (l.input, (l.x, l.y, l.w, l.h, l.zorder_offset)))
-        .collect();
-
-    for i in 0..num_inputs {
-        let pad_name = format!("sink_{}", pad_base + i);
-        let Some(pad) = find_pad(compositor, &pad_name) else {
-            continue;
-        };
-        // Clear any lingering control bindings from a previous fade — otherwise
-        // they keep driving the property and our set_property calls below would
-        // be invisible until the next take rebuilds bindings.
-        for prop in ["alpha", "xpos", "ypos", "width", "height", "zorder"] {
-            if let Some(binding) = pad.control_binding(prop) {
-                pad.remove_control_binding(&binding);
-            }
-        }
-        if Some(i) == bg {
-            let aspect = effective_source_aspect(
-                src_aspects.get(&i).copied().unwrap_or(fallback_aspect),
-                transforms.get(&i),
-            );
-            let (x, y, w, h) = aspect_fit_rect(rx, ry, rw, rh, aspect);
-            pad.set_property("xpos", x);
-            pad.set_property("ypos", y);
-            pad.set_property("width", w);
-            pad.set_property("height", h);
-            pad.set_property("alpha", 1.0f64);
-            pad.set_property("zorder", bg_zorder);
-        } else if let Some(&(ox, oy, ow, oh, off)) = layout_map.get(&i) {
-            pad.set_property("xpos", ox);
-            pad.set_property("ypos", oy);
-            pad.set_property("width", ow);
-            pad.set_property("height", oh);
-            pad.set_property("alpha", 1.0f64);
-            pad.set_property("zorder", overlay_zorder + off);
-        } else {
-            pad.set_property("alpha", 0.0f64);
-        }
-        // Crop applies to the source wherever it renders in this PiP —
-        // including hidden pads, whose retained transform stays staged so the
-        // source returns pre-framed. Non-PiP reveal paths wipe crop
-        // themselves (apply_input_group_to_region, the classic-take reset).
-        let crop = transforms.get(&i).copied().unwrap_or_default();
-        set_pad_crop(&pad, &crop);
     }
 }
